Update Dependabot and Gradle files for Dependabot
This commit lays the foundation for dependabot to better manage
version updates in the future. Specifically, there were 3
issues addressed in this PR:
- Gradle build files were using set() instead of = which
causes dependabot to miss upgrading the dependency
- Gradle build files were specifying the version of
io.spring.dpeendency-management which necessitated logic
in the dependabot config, which was not working correctly
- Dependabot was not using groups to limit the number of
PRs sent for a given version update

After this update Dependabot should be able to ignore all
patch versions going forward for Vaadin.

This commit also updates the Spring Boot and Vaadin versions
to the latest patch releases.